db connection happens only once
