+++ main.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"database/sql"
+	"os"
 	"time"
 
 	"github.com/sirupsen/logrus"
@@ -12,12 +14,36 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 )
 
-func handler(w http.ResponseWriter, r *http.Request) {
+type CollectorApp struct {
+	Database *sql.DB
+}
+
+func connectToDB() (*sql.DB, error) {
+	DBUsername := os.Getenv("DB_USER")
+	DBPassword := os.Getenv("DB_PASSWORD")
+	DBURL := os.Getenv("DB_URL")
+	DBPort := os.Getenv("DB_PORT")
+
+	DBConnStr := DBUsername + ":" + DBPassword + "@tcp(" + DBURL + ":" + DBPort + ")/"
+	db, err := sql.Open("mysql", DBConnStr)
+	if err != nil {
+		return nil, err
+	}
+
+	err = db.Ping()
+	if err != nil {
+		return nil, err
+	}
+
+	return db, nil
+}
+
+func (collector *CollectorApp) handler(w http.ResponseWriter, r *http.Request) {
 	switch r.Method {
 	case http.MethodGet:
-		actions.ResultsHandler(w, r)
+		actions.ResultsHandler(w, r, collector.Database)
 	case http.MethodPost:
-		actions.ParserHandler(w, r)
+		actions.ParserHandler(w, r, collector.Database)
 	default:
 		_, writeErr := w.Write([]byte(actions.InvalidRequestErr + "\n"))
 		if writeErr != nil {
@@ -28,7 +54,12 @@ func handler(w http.ResponseWriter, r *http.Request) {
 }
 
 func main() {
-	http.HandleFunc("/", handler)
+	// connect to DB
+	db, _ := connectToDB()
+
+	collector := &CollectorApp{Database: db}
+
+	http.HandleFunc("/", collector.handler)
 	server := &http.Server{
 		Addr:         ":8080",
 		ReadTimeout:  10 * time.Second,
@@ -36,6 +67,8 @@ func main() {
 	}
 	err := server.ListenAndServe()
 	if err != nil {
+		db.Close()
 		log.Fatal(err)
 	}
+	db.Close()
 }
